+++ tf-modules/data-persistence/postgres_migration.tf
@@ -7,4 +7,5 @@ module "db_migration" {
   tags                       = merge(var.tags, { Deployment = var.prefix })
   vpc_id                     = var.vpc_id
   rds_security_group_id      = var.rds_security_group_id
+  lambda_timeout             = var.db_migration_lambda_timeout
 }